# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "API Concesionarios",
    "version": "1.0.2"
  },
  "definitions": {
    "concesionarios": {
      "type": "object",
      "properties": {
        "nombre": { "type": "string" },
        "direccion": { "type": "string" },
        "listCoches": {
          "type": "array",
          "items": { "$ref": "#/definitions/coches" }
        }
      }
    },
    "coches": {
      "type": "object",
      "properties": {
        "marca": { "type": "string" },
        "modelo": { "type": "string" }
      }
    }
  },

  "paths": {
    "/concesionarios": {
      "get": {
        "summary": "get all concesionarios",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/concesionarios" }
            }
          }
        }
      },
      "post": {
        "summary": "add a new concesionario",
        "parameters": [
          {
            "name": "concesionarios",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/concesionarios" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/concesionarios/{id}": {
      "get": {
        "summary": "Get Concesionario by id",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/concesionarios" }
          }
        }
      },

      "put": {
        "summary": "update a concesionario by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "concesionarios",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/concesionarios" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "summary": "Delete a concesionario by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/concesionarios/{id}/coches": {
      "get": {
        "summary": "get all coches de un concesionario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "503": { "description": "Service Unavailable" }
        }
      }
    }
  }
}